Remove git checkout from CommandRunner's reponsibility
