+++ src/main/java/reposense/git/GitChecker.java
@@ -1,11 +1,15 @@
 package reposense.git;
 
+import static reposense.system.CommandRunner.runCommand;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Date;
 import java.util.logging.Logger;
 
 import reposense.commits.model.CommitResult;
-import reposense.system.CommandRunner;
 import reposense.system.LogsManager;
+
 
 public class GitChecker {
 
@@ -24,8 +28,9 @@ public class GitChecker {
         checkout(root, commit.getHash());
     }
 
-    public static void checkout(String root, String commitHash) {
-        CommandRunner.checkout(root, commitHash);
+    public static void checkout(String root, String hash) {
+        Path rootPath = Paths.get(root);
+        runCommand(rootPath, "git checkout " + hash);
     }
 
     /**
@@ -34,7 +39,20 @@ public class GitChecker {
      * @throws CommitNotFoundException if commits before {@code untilDate} cannot be found.
      */
     public static void checkoutToDate(String root, String branchName, Date untilDate) throws CommitNotFoundException {
-        CommandRunner.checkoutToDate(root, branchName, untilDate);
+        if (untilDate == null) {
+            return;
+        }
+
+        Path rootPath = Paths.get(root);
+
+        String substituteCommand = "git rev-list -1 --before="
+                + Util.GIT_LOG_UNTIL_DATE_FORMAT.format(untilDate) + " " + branchName;
+        String hash = runCommand(rootPath, substituteCommand);
+        if (hash.isEmpty()) {
+            throw new CommitNotFoundException("Commit before until date is not found.");
+        }
+        String checkoutCommand = "git checkout " + hash;
+        runCommand(rootPath, checkoutCommand);
     }
 }
